requirements.txt
Django==1.7
Markdown==2.5.1
dj-database-url==0.3.0
django-filter==0.8
django-grappelli==2.6.2
djangorestframework==2.4.3
ipdb==0.8
ipython==2.3.0
